+++ scaleph-engine/scaleph-engine-flink-kubernetes/src/main/java/cn/sliew/scaleph/engine/flink/kubernetes/resource/definition/job/FlinkDeploymentJobConverter.java
@@ -18,6 +18,7 @@
 
 package cn.sliew.scaleph.engine.flink.kubernetes.resource.definition.job;
 
+import cn.sliew.milky.common.util.JacksonUtil;
 import cn.sliew.scaleph.common.util.SeaTunnelReleaseUtil;
 import cn.sliew.scaleph.config.resource.ResourceLabels;
 import cn.sliew.scaleph.config.resource.ResourceNames;
@@ -26,23 +27,35 @@ import cn.sliew.scaleph.dao.entity.master.ws.WsFlinkArtifactJar;
 import cn.sliew.scaleph.dao.entity.master.ws.WsFlinkArtifactSql;
 import cn.sliew.scaleph.engine.flink.kubernetes.operator.spec.FlinkDeploymentSpec;
 import cn.sliew.scaleph.engine.flink.kubernetes.operator.spec.JobSpec;
-import cn.sliew.scaleph.engine.flink.kubernetes.resource.handler.FileFetcherFactory;
 import cn.sliew.scaleph.engine.flink.kubernetes.resource.definition.deployment.FlinkDeployment;
 import cn.sliew.scaleph.engine.flink.kubernetes.resource.definition.deployment.FlinkDeploymentConverter;
+import cn.sliew.scaleph.engine.flink.kubernetes.resource.handler.FileFetcherFactory;
 import cn.sliew.scaleph.engine.flink.kubernetes.service.dto.WsFlinkKubernetesJobDTO;
+import cn.sliew.scaleph.engine.seatunnel.service.SeatunnelConfigService;
+import cn.sliew.scaleph.engine.seatunnel.service.WsDiJobService;
+import cn.sliew.scaleph.engine.seatunnel.service.dto.WsDiJobDTO;
 import cn.sliew.scaleph.kubernetes.resource.ResourceConverter;
 import io.fabric8.kubernetes.api.model.ObjectMetaBuilder;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 
-public enum FlinkDeploymentJobConverter implements ResourceConverter<WsFlinkKubernetesJobDTO, FlinkDeploymentJob> {
-    INSTANCE;
+@Component
+public class FlinkDeploymentJobConverter implements ResourceConverter<WsFlinkKubernetesJobDTO, FlinkDeploymentJob> {
+
+    @Autowired
+    private SeatunnelConfigService seatunnelConfigService;
+    @Autowired
+    private WsDiJobService wsDiJobService;
+    @Autowired
+    private FileFetcherFactory fileFetcherFactory;
 
     @Override
-    public FlinkDeploymentJob convertTo(WsFlinkKubernetesJobDTO source) {
+    public FlinkDeploymentJob convertTo(WsFlinkKubernetesJobDTO source) throws Exception {
         FlinkDeploymentJob deployment = new FlinkDeploymentJob();
         FlinkDeployment flinkDeployment = FlinkDeploymentConverter.INSTANCE.convertTo(source.getFlinkDeployment());
         ObjectMetaBuilder builder = new ObjectMetaBuilder(flinkDeployment.getMetadata(), true);
@@ -59,7 +72,7 @@ public enum FlinkDeploymentJobConverter implements ResourceConverter<WsFlinkKube
             jobSpec.setEntryClass(flinkArtifactJar.getEntryClass());
             jobSpec.setArgs(StringUtils.split(flinkArtifactJar.getJarParams(), " "));
             spec.setJob(jobSpec);
-            FileFetcherFactory.INSTANCE.customize(source, deployment);
+            fileFetcherFactory.customize(source, deployment);
         }
         if (source.getFlinkArtifactSql() != null) {
             WsFlinkArtifactSql flinkArtifactSql = source.getFlinkArtifactSql();
@@ -75,7 +88,9 @@ public enum FlinkDeploymentJobConverter implements ResourceConverter<WsFlinkKube
             JobSpec jobSpec = new JobSpec();
             jobSpec.setJarURI(ResourceNames.LOCAL_SCHEMA + "/opt/seatunnel/starter/" + SeaTunnelReleaseUtil.STARTER_JAR_NAME);
             jobSpec.setEntryClass(SeaTunnelReleaseUtil.SEATUNNEL_MAIN_CLASS);
-            List<String> args = Arrays.asList("--config", "todo config");
+            WsDiJobDTO wsDiJobDTO = wsDiJobService.queryJobGraph(wsDiJob.getId());
+            String config = seatunnelConfigService.buildConfig(wsDiJobDTO);
+            List<String> args = Arrays.asList("--config", JacksonUtil.toJsonNode(config).toString());
             jobSpec.setArgs(args.toArray(new String[2]));
             spec.setJob(jobSpec);
         }
