+++ scaleph-engine/scaleph-engine-flink-kubernetes/src/main/java/cn/sliew/scaleph/engine/flink/kubernetes/service/impl/WsFlinkKubernetesJobServiceImpl.java
@@ -18,11 +18,9 @@
 
 package cn.sliew.scaleph.engine.flink.kubernetes.service.impl;
 
-import cn.sliew.milky.common.util.JacksonUtil;
 import cn.sliew.scaleph.common.util.UUIDUtil;
 import cn.sliew.scaleph.dao.entity.master.ws.WsFlinkKubernetesJob;
 import cn.sliew.scaleph.dao.mapper.master.ws.WsFlinkKubernetesJobMapper;
-import cn.sliew.scaleph.engine.flink.kubernetes.resource.definition.job.FlinkDeploymentJob;
 import cn.sliew.scaleph.engine.flink.kubernetes.resource.definition.job.FlinkDeploymentJobConverter;
 import cn.sliew.scaleph.engine.flink.kubernetes.resource.definition.job.FlinkSessionJobConverter;
 import cn.sliew.scaleph.engine.flink.kubernetes.service.FlinkKubernetesOperatorService;
@@ -34,13 +32,11 @@ import cn.sliew.scaleph.engine.flink.kubernetes.service.param.WsFlinkKubernetesJ
 import cn.sliew.scaleph.engine.flink.kubernetes.service.param.WsFlinkKubernetesJobUpdateParam;
 import cn.sliew.scaleph.engine.seatunnel.service.SeatunnelConfigService;
 import cn.sliew.scaleph.engine.seatunnel.service.WsDiJobService;
-import cn.sliew.scaleph.engine.seatunnel.service.dto.WsDiJobDTO;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.Arrays;
 import java.util.List;
 
 import static cn.sliew.milky.common.check.Ensures.checkState;
@@ -52,10 +48,9 @@ public class WsFlinkKubernetesJobServiceImpl implements WsFlinkKubernetesJobServ
     private WsFlinkKubernetesJobMapper wsFlinkKubernetesJobMapper;
     @Autowired
     private FlinkKubernetesOperatorService flinkKubernetesOperatorService;
+
     @Autowired
-    private SeatunnelConfigService seatunnelConfigService;
-    @Autowired
-    private WsDiJobService wsDiJobService;
+    private FlinkDeploymentJobConverter flinkDeploymentJobConverter;
 
     @Override
     public Page<WsFlinkKubernetesJobDTO> list(WsFlinkKubernetesJobListParam param) {
@@ -79,15 +74,7 @@ public class WsFlinkKubernetesJobServiceImpl implements WsFlinkKubernetesJobServ
         WsFlinkKubernetesJobDTO wsFlinkKubernetesJobDTO = selectOne(id);
         switch (wsFlinkKubernetesJobDTO.getDeploymentKind()) {
             case FLINK_DEPLOYMENT:
-                FlinkDeploymentJob flinkDeploymentJob = FlinkDeploymentJobConverter.INSTANCE.convertTo(wsFlinkKubernetesJobDTO);
-                if (wsFlinkKubernetesJobDTO.getWsDiJob() != null) {
-                    WsDiJobDTO wsDiJobDTO = wsDiJobService.queryJobGraph(wsFlinkKubernetesJobDTO.getWsDiJob().getId());
-                    wsDiJobDTO.getWsFlinkArtifact().setName(wsFlinkKubernetesJobDTO.getJobId());
-                    String config = seatunnelConfigService.buildConfig(wsDiJobDTO);
-                    List<String> args = Arrays.asList("--config", JacksonUtil.toJsonNode(config).toString());
-                    flinkDeploymentJob.getSpec().getJob().setArgs(args.toArray(new String[2]));
-                }
-                return flinkDeploymentJob;
+                return flinkDeploymentJobConverter.convertTo(wsFlinkKubernetesJobDTO);
             case FLINK_SESSION_JOB:
                 return FlinkSessionJobConverter.INSTANCE.convertTo(wsFlinkKubernetesJobDTO);
             default:
